+++ classes/prefix_core_WPadmin/class.prefix_core_WPadmin.php
@@ -4,7 +4,7 @@
  *
  * Backend area to manage configuration file
  * Author:      David Voglgsnag
- * @version     1.3.1
+ * @version     1.3.2
  *
  */